improve variable matcher

+++ bug-definitions/src/test/java/org/bugby/pattern/example/test/VariableAssignCheck2.java
@@ -3,7 +3,7 @@ package org.bugby.pattern.example.test;
 public class VariableAssignCheck2 {
 
 	public void method2() {
-		double x = 10;
+		short x = 10;
 		if (x == 2) {
 			return;
 		}
